requirements.txt
torch --index-url https://download.pytorch.org/whl/cu121
aiogram
pytesseract
openai-whisper
python-dotenv